requirement.txt
Flask==2.0.1
gunicorn==20.1.0
pylint==2.10.2
python-dotenv==0.19.0
requests==2.22.0

